Cleanup: move getWorkspaceRoot to MayaUtilities
